A whole lot of cleanup and extra functionality.
This code is now ready to run on an ongoing basis.

It does the right thing when run repeatedly against the same logfile,
it can run against different files (i.e. you can point it at a different
file to process), in such a way that you can run it from a logrotate
postrotate script to capture the last dregs of the old logfile, and it
now has some useful stats reporting and stuff.